Core: avoid signed overflow when assembling DNS TTLs
Four places in the resolver build a 32-bit value out of four bytes of a
DNS response without making the arithmetic unsigned:

    ttl = (an->ttl[0] << 24) + (an->ttl[1] << 16)
        + (an->ttl[2] << 8) + (an->ttl[3]);

an->ttl[0] is a u_char and promotes to int, so any TTL with the top bit
set shifts a value of at least 128 left by 24 places, which does not
fit in int.  The A record address is assembled the same way, so any
address in 128.0.0.0/1 reaches it as well.

gcc and clang wrap, so the value comes out negative and the ttl < 0
guard below catches it.  That guard is the reason to fix this: the only
way ttl can be negative is through the overflow, so value-range
propagation is entitled to conclude ttl >= 0 and drop the check.  The
value flows on into rn->valid as a cache lifetime, where a TTL of
0x80000000 would be kept as roughly 68 years.

Make the arithmetic unsigned, as ngx_inet.c, ngx_stream_geo_module.c
and ngx_http_grpc_module.c already do.  Converting back to int32_t is
implementation-defined rather than undefined, and both compilers define
it as the wrap that happens today, so the guard keeps firing on exactly
the same inputs and observable behaviour is unchanged.

Found with a libFuzzer harness driving ngx_resolver_process_response()
under UBSan.

+++ src/core/ngx_resolver.c
@@ -2194,8 +2194,10 @@ ngx_resolver_process_a(ngx_resolver_t *r, u_char *buf, size_t n,
         type = (an->type_hi << 8) + an->type_lo;
         class = (an->class_hi << 8) + an->class_lo;
         len = (an->len_hi << 8) + an->len_lo;
-        ttl = (an->ttl[0] << 24) + (an->ttl[1] << 16)
-            + (an->ttl[2] << 8) + (an->ttl[3]);
+        ttl = (int32_t) (((uint32_t) an->ttl[0] << 24)
+                         + ((uint32_t) an->ttl[1] << 16)
+                         + ((uint32_t) an->ttl[2] << 8)
+                         + an->ttl[3]);
 
         if (class != 1) {
             ngx_log_error(r->log_level, r->log, 0,
@@ -2356,8 +2358,10 @@ ngx_resolver_process_a(ngx_resolver_t *r, u_char *buf, size_t n,
 
             if (type == NGX_RESOLVE_A) {
 
-                addr[j] = htonl((buf[i] << 24) + (buf[i + 1] << 16)
-                                + (buf[i + 2] << 8) + (buf[i + 3]));
+                addr[j] = htonl(((uint32_t) buf[i] << 24)
+                                + ((uint32_t) buf[i + 1] << 16)
+                                + ((uint32_t) buf[i + 2] << 8)
+                                + buf[i + 3]);
 
                 if (++j == naddrs) {
 
@@ -2736,8 +2740,10 @@ ngx_resolver_process_srv(ngx_resolver_t *r, u_char *buf, size_t n,
         type = (an->type_hi << 8) + an->type_lo;
         class = (an->class_hi << 8) + an->class_lo;
         len = (an->len_hi << 8) + an->len_lo;
-        ttl = (an->ttl[0] << 24) + (an->ttl[1] << 16)
-            + (an->ttl[2] << 8) + (an->ttl[3]);
+        ttl = (int32_t) (((uint32_t) an->ttl[0] << 24)
+                         + ((uint32_t) an->ttl[1] << 16)
+                         + ((uint32_t) an->ttl[2] << 8)
+                         + an->ttl[3]);
 
         if (class != 1) {
             ngx_log_error(r->log_level, r->log, 0,
@@ -3301,8 +3307,10 @@ valid:
         type = (an->type_hi << 8) + an->type_lo;
         class = (an->class_hi << 8) + an->class_lo;
         len = (an->len_hi << 8) + an->len_lo;
-        ttl = (an->ttl[0] << 24) + (an->ttl[1] << 16)
-            + (an->ttl[2] << 8) + (an->ttl[3]);
+        ttl = (int32_t) (((uint32_t) an->ttl[0] << 24)
+                         + ((uint32_t) an->ttl[1] << 16)
+                         + ((uint32_t) an->ttl[2] << 8)
+                         + an->ttl[3]);
 
         if (class != 1) {
             ngx_log_error(r->log_level, r->log, 0,
